Extract fixture data out of the <MyBids> component
